Session 13 Finalizado

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUseServerSession } from "@/auth/auth-actions";
 import prisma from "@/lib/prisma";
 import { NextResponse, NextRequest } from "next/server";
 import { boolean, object, string } from "yup";
@@ -36,12 +37,16 @@ const postSchema = object({
 });
 
 export async function POST(request: Request) {
+    const user = await getUseServerSession();
+
+    if (!user) return NextResponse.json("No autorizado", { status: 401 });
+
     try {
         const { description, complete } = await postSchema.validate(
             await request.json()
         );
         const todo = await prisma.todo.create({
-            data: { description, complete },
+            data: { description, complete, userId: user.id },
         });
         return NextResponse.json(todo);
     } catch (error) {
@@ -50,8 +55,12 @@ export async function POST(request: Request) {
 }
 
 export async function DELETE(req: Request) {
+    const user = await getUseServerSession();
+
+    if (!user) return NextResponse.json("No autorizado", { status: 401 });
+
     const deleteTodo = await prisma.todo.deleteMany({
-        where: { complete: true },
+        where: { complete: true, userId: user.id },
     });
 
     return NextResponse.json(deleteTodo);
